single component for space features list
